+++ NEWS.md
@@ -1,3 +1,35 @@
+# hint 1.65.0
+
+* Update error report object in backend include version info 
+
+# hint 1.64.0
+
+* Improve error reporting process
+
+# hint 1.63.0
+
+* Display warning alert at bottom of Model Options step
+
+# hint 1.62.1
+
+* ADR upload fails if download attempted while in progress
+
+# hint 1.62.0
+
+* Upload release to ADR epic
+
+# hint 1.61.0
+
+* Include WarningAlert in Stepper
+
+# hint 1.60.0
+
+* Do not poll for ADR dataset updates when using a releaser
+
+# hint 1.59.0
+
+* Do not auto-progress to calibrate step if there are model fit warnings
+
 # hint 1.58.0
 
 * Endpoint for posting error report to Teams